Move directories to match new scheme
