+++ decapodes/configuration/decapodes.it
@@ -1,42 +1,22 @@
-struct Dimensionality
-	surface::Int64
-	embedding::Int64
+@sum MeshValue begin
+  File(uri::String, checksum::String, format::String, shape::Vector{Int64})
+  Values(values::Vector{Int64})
 end
 
-@sum Region begin
-	Equation(equation::String)
-end
-
-struct SubmeshRelation
-  parent::String
-  child::String
-end
-
-struct Mesh
-  id::String
-  description::String
-  file::String
-  checksum::String
-  submesh_relation::Vector{SubmeshRelation}
-  dimensionality::Dimensionality
-  vertex_count::Int64
-  regions::Vector{Region}
-end
-
-struct Configuration
-  meshes::Vector{Mesh}
-  dimensionality::Int64
-  boundary_mesh::String
+struct Constants
+  name::String
+  global_ref::String
 end
 
 struct Header
-	id::String
-    description::String
-    name::String
-    parent::String
+  id::String
+  description::String
+  name::String
+  parent::String
 end
 
-struct ASKEMDecapodeConfig
+struct ASKEMDecapodeParameterization
   header::Header
-  configuration::Configuration
+  parameters::Vector{Float64}
+  constants::Vector{Constants}
 end
