--- decapodes/parameterization/decapodes_schema.json
@@ -1,126 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "$defs": {
-    "File": {
-      "type": "object",
-      "properties": {
-        "tag": {
-          "const": "File"
-        },
-        "uri": {
-          "type": "string",
-          "$comment": "Str"
-        },
-        "checksum": {
-          "type": "string",
-          "$comment": "Str"
-        },
-        "format": {
-          "type": "string",
-          "$comment": "Str"
-        },
-        "shape": {
-          "type": "array",
-          "items": {
-            "type": "string",
-            "$comment": "I64"
-          }
-        }
-      },
-      "required": [
-        "uri",
-        "checksum",
-        "format",
-        "shape"
-      ]
-    },
-    "Values": {
-      "type": "object",
-      "properties": {
-        "tag": {
-          "const": "Values"
-        },
-        "values": {
-          "type": "array",
-          "items": {
-            "type": "string",
-            "$comment": "I64"
-          }
-        }
-      },
-      "required": [
-        "values"
-      ]
-    },
-    "Constants": {
-      "type": "object",
-      "properties": {
-        "name": {
-          "type": "string",
-          "$comment": "Str"
-        },
-        "global_ref": {
-          "type": "string",
-          "$comment": "Str"
-        }
-      },
-      "required": [
-        "name",
-        "global_ref"
-      ]
-    },
-    "Header": {
-      "type": "object",
-      "properties": {
-        "id": {
-          "type": "string",
-          "$comment": "Str"
-        },
-        "description": {
-          "type": "string",
-          "$comment": "Str"
-        },
-        "name": {
-          "type": "string",
-          "$comment": "Str"
-        },
-        "parent": {
-          "type": "string",
-          "$comment": "Str"
-        }
-      },
-      "required": [
-        "id",
-        "description",
-        "name",
-        "parent"
-      ]
-    },
-    "ASKEMDecapodeParameterization": {
-      "type": "object",
-      "properties": {
-        "header": {
-          "$ref": "#/$defs/Header"
-        },
-        "parameters": {
-          "type": "array",
-          "items": {
-            "type": "number",
-            "$comment": "F64"
-          }
-        },
-        "constants": {
-          "type": "array",
-          "items": {
-            "$ref": "#/$defs/Constants"
-          }
-        }
-      },
-      "required": [
-        "header",
-        "parameters",
-        "constants"
-      ]
-    }
-  }
-}
